+++ packages/react-bindings/src/contexts/StreamI18nContext.tsx
@@ -3,11 +3,13 @@ import {
   PropsWithChildren,
   useContext,
   useEffect,
+  useMemo,
   useState,
 } from 'react';
 import {
   defaultTranslationFunction,
   StreamI18n,
+  TranslationLanguage,
   TranslationsMap,
 } from '../i18n';
 
@@ -22,8 +24,8 @@ const StreamI18nContext = createContext<StreamI18nContextValue>({
 
 type CreateI18nParams = {
   i18nInstance?: StreamI18n;
-  language?: string;
-  fallbackLanguage?: string;
+  language?: TranslationLanguage;
+  fallbackLanguage?: TranslationLanguage;
   translationsOverrides?: TranslationsMap;
 };
 
@@ -33,16 +35,15 @@ export const StreamI18nProvider = ({
   children,
   ...createI18nParams
 }: PropsWithChildren<StreamI18nProviderProps>) => {
-  const { i18n, t } = useCreateI18n(createI18nParams);
-
+  const api = useCreateI18n(createI18nParams);
   return (
-    <StreamI18nContext.Provider value={{ t, i18n }}>
+    <StreamI18nContext.Provider value={api}>
       {children}
     </StreamI18nContext.Provider>
   );
 };
 
-export const useCreateI18n = ({
+const useCreateI18n = ({
   i18nInstance,
   language,
   fallbackLanguage,
@@ -57,24 +58,24 @@ export const useCreateI18n = ({
         translationsOverrides,
       }),
   );
-  const [t, setTranslationFn] = useState<StreamI18n['t']>(
-    () => defaultTranslationFunction,
-  );
-
+  const [t, setTranslationFn] = useState<StreamI18n['t']>(() => i18n.t);
   useEffect(() => {
-    const { isInitialized } = i18n;
-    if (!isInitialized) {
-      i18n.init().then((_i18n) => setTranslationFn(() => _i18n.i18nInstance.t));
-      return;
+    if (!i18n.isInitialized) {
+      i18n.init().then(() => setTranslationFn(() => i18n.t));
+    } else if (i18n.currentLanguage !== language) {
+      i18n.changeLanguage(language).then(() => setTranslationFn(() => i18n.t));
     }
-    if (language && i18n?.currentLanguage !== language) {
-      i18n.changeLanguage(language).catch((err) => {
-        console.log('Error while changing language', err);
-      });
-    }
-  }, [i18n, i18nInstance, language, translationsOverrides]);
+  }, [i18n, language]);
 
-  return { i18n, t };
+  return useMemo(() => ({ i18n, t }), [i18n, t]);
 };
 
+/**
+ * A hook to get the i18n instance and translation function from the closest provider.
+ *
+ * Example usage:
+ * const { t, i18n } = useI18n();
+ * const message = t('hello_world');
+ * console.log(message);
+ */
 export const useI18n = () => useContext(StreamI18nContext);